+++ eslint.config.mjs
@@ -76,6 +76,10 @@ export default defineConfig(
       "@typescript-eslint/no-unsafe-call": "off",
       "@typescript-eslint/no-unsafe-member-access": "off",
       "@typescript-eslint/no-unsafe-return": "off",
+      "@typescript-eslint/no-unused-vars": [
+        "error",
+        { varsIgnorePattern: "^_" },
+      ],
       "@typescript-eslint/unbound-method": "off",
 
       // these rules are slow, require extensive config, and/or don't provide much