+++ packages/config/src/parse.test.ts
@@ -99,7 +99,6 @@ describe("readConfig", () => {
       "pnpm-workspace.yaml": "packages: ['packages/*']",
     });
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { ["$schema" as never]: _, ...rest } = defaultConfig;
     const result = await readConfig(cwd);
     expect(result.config).toStrictEqual(rest);
@@ -324,6 +323,12 @@ describe("validateConfig", () => {
         expected: { ...defaultConfig, ignore: ["pkg-a", "@pkg/a", "@pkg/b"] },
       },
       {
+        name: "ignore: allows unmatched globs",
+        pkgs: ["pkg-a", "pkg-b"],
+        config: { ignore: ["not-a-valid-package"] },
+        expected: { ...defaultConfig, ignore: [] },
+      },
+      {
         name: "privatePackages: false",
         config: { privatePackages: false },
         expected: {
@@ -521,18 +526,6 @@ describe("validateConfig", () => {
         name: "ignore: array of non-string",
         config: { ignore: [123, "pkg-a"] },
         errors: ["Expected string"],
-      },
-      {
-        name: "rule: ignoredPatternsExist",
-        pkgs: [],
-        config: { ignore: ["pkg-a"] },
-        errors: ['Invalid path: The package or glob "pkg-a" does not match'],
-      },
-      {
-        name: "rule: ignoredPatternsExist: glob",
-        pkgs: [],
-        config: { ignore: ["pkg-*"] },
-        errors: ['Invalid path: The package or glob "pkg-*" does not match'],
       },
       // privatePackages
       {